--- src/app/pages/navigation/common/service/countries.service.ts
@@ -1,14 +1,0 @@
-import {Injectable} from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-
-export class CountriesService {
-
-  constructor() {
-  }
-
-  public getIPAddress(): any {
-  }
-}
