+++ src/app/pages/navigation/contact/contact.component.ts
@@ -1,14 +1,11 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {COUNTRIES_EN} from '../common/countries/data/Countries_EN';
 import {ICountry} from '../common/countries/data/ICountry';
 import {TranslateComponent} from '../../../core/translate/translate.component';
 import {TranslateWrapperService} from '../../../core/service/translate-wrapper.service';
 import {LangChangeEvent, TranslateService} from '@ngx-translate/core';
-import {LanguageEnum} from '../../../core/enum/LanguageEnum';
-import {COUNTRIES_ES} from '../common/countries/data/Countries_ES';
 import {IpService} from '../../../core/service/ip.service';
-import {ISite} from '../../private/sites/data/ISite';
+import {CountriesService} from '../common/countries/countries.service';
 
 @Component({
   selector: 'ambm-contact',
@@ -25,14 +22,15 @@ export class ContactComponent extends TranslateComponent implements OnInit {
   constructor(private formBuilder: FormBuilder,
               public translate: TranslateService,
               public translateWrapperService: TranslateWrapperService,
-              private ipService: IpService) {
+              private ipService: IpService,
+              private countriesService: CountriesService) {
     super(translate, translateWrapperService);
     this.onLangChange();
     this.validateFormBuilder();
   }
 
   ngOnInit(): void {
-    this.getCountries(this.translate.currentLang);
+    this.countries = this.countriesService.getCountries(this.translate.currentLang);
   }
 
   private validateFormBuilder(): void {
@@ -78,12 +76,6 @@ export class ContactComponent extends TranslateComponent implements OnInit {
     this.validateForm.reset();
   }
 
-  searchCountry(id: string): ICountry {
-    return this.countries.filter(country => {
-      return country.iata.toLowerCase().includes(id);
-    })[0];
-  }
-
   changeCountry(country: ICountry): void {
     this.countrySelected = country;
   }
@@ -91,19 +83,11 @@ export class ContactComponent extends TranslateComponent implements OnInit {
 
   private onLangChange(): void {
     this.translate.onLangChange.subscribe((params: LangChangeEvent) => {
-      this.getCountries(params.lang);
+      this.countries = this.countriesService.getCountries(params.lang);
       if (this.country) {
-        this.country = this.searchCountry(this.country.iata);
+        this.country = this.countriesService.searchCountry(this.country.iata);
       }
     });
-  }
-
-  getCountries(language: string): void {
-    if (language === LanguageEnum.ENGLISH) {
-      this.countries = COUNTRIES_EN;
-    } else {
-      this.countries = COUNTRIES_ES;
-    }
   }
 
 }
